This changes variable names to be standardized...
It fixes ALOT of problems with malformed XML configurations
sprinkled throughout are quite a few bug fixes and console log cleanups
this makes way for other "modes" of questor... and fixes some issues with running questor in the dedicated "salvage" mode.

Eventually I'll break this commit up into pieces and commit to experimental.