remove unnecessary console log spam every iteration... this was NOT actually approaching every second.
the issue was it was checking the nextorbit timer correctly, and only spamming this message to the console log after the timer were up and we COULD approach again, but we were already approaching so the command to actually approach was not run, thus the nextapproach timestamp not updated. the best fix was to simply remove this unnecessary console log line. If we actually approach we say so in the log.