Last round of fixes... tomorrow Ill start a new branch and merge in all these changes, hopefully with all the file history intact. This branch managed to kill the file history on a few files when moving things around.
Ive disabled the target following in the GUI stuff as it was unstable. The detail available there was unprecedented, but the values also seemed 'wrong' at times. I'm not sure if it was the speed at which updates were being pulled or something else. After the merge I may play with it again to see it I can get it to behave without crashing innerspace.