FIX drones - evidentially WHEN SCRAMBLED keeping the drones out while they are taking shield damage but pulling them when they go into armor does not work.  I reverted that behavior change.
I did however fix the "mission is done" so pull the drones regardless of if we are scrambled or not problem. If the mission is marked as done but you are still scrambled the drones will stay out and shoot things until you are not scrambled anymore.

And we have better console logging of actual number of drones in space when we engage targets, etc