This adds detailed CurrentTarget health Statistics to the console log. This defaults to off.
If you want it on you need to add "<detailedCurrentTargetHealthLogging>true</detailedCurrentTargetHealthLogging>"
to your characters settings xml file.

This also fixes a longstanding annoyance where the combat module would switch off an almost dead NPC because a higher priority one became available (for whatever reason). This changes that behavior. If the currenttarget is less than 60% armor it will continue to kill the currenttarget until it is dead.