New debugging settings that add more logging - you will want these off unless you know what you are looking for specifically.
I also added useInnerspace. If this is false all innerspace related commands will not be used. keep in mind that we use innerspace heavily to restart questor after stopping for high memory usage, or getting stuck in a mission and wanting to restart to recover, etc. If you wish to use this feature I hope you know how to program in x# and can fix the restarting questor code to work without innerspace.