Add DebuggingWait Action that is now available in mission XMLs. It will wait a specified time based on the timeout parameter or wait 1200 seconds if no parameter is specified
This is not a good idea to use on any 'working' mission XML. It is nice however to add to a mission XML that you want to watch the questor complete for whatever reason and you arent quite sure when it will pull that mission. This at least allow you to have a much higher chance of seeing what is happening.