Fixed for the CombatHelper behavior.
If you specify what damage type you want to load you can use the Arm state to load ammo from the configured ammohangar
unloadloot will unload anything in the cargo to the configured loothangar
CombatHelper no longer waits for the NPCs to target it before shooting... it will initiate combat, thus allowing for your 'main' and your 'alt' to be in the same mission pocket (or belt or wherever) and the combat helper will shoot things even if it has no aggro. Drones are the same, they will now aggress even without aggro.
The fittingmanager feature cannot be used with CombatHelper (and I do not see a logical reason to want it when you are at the keyboard)

I also added a bit of logging that will likely be moved to a dedicated behavior later.