When logging in while in space questor was throwing exceptions opening the cargohold and occasionally pulling the agent information.
make sure the cargohold is opened in space before we attempt to determine our range as we need to know what ammo we have.

There is probably an underlying cause to this but for now lets handle the exceptions and assume they will resolve themselves after eve has time to settle in.