Add new setting
<orbitStructure>false</orbitStructure>
for speed tanks. If true it will orbit the "gate" if there is one. Otherwise it will orbit the target as usual.
If this is set to false, or not set at all, the old behavior is preserved.

This was save your ship from getting hung on things... and SHOULD speed up missions as you wont get so far from the gate.