Fixes: https://github.com/ISeeDEDPpl/Questor/issues/106
Before this if their was a LARGE gap between eve starting and questor logging in because of the schedules.xml upon login it would immediately schedule a restart of questor... this resolves that.

That being said, why have a large gap between starting questor and having it login? use windows scheduler to start eve right before you want it to be running ffs.