Add debug logging to salvage.cs
If you wish to use this you will need to set <DebugSalvage>true</DebugSalvage>
in your characters settings xml file.

keep in mind that this is going to be horribly chatty so you don't want to leave this debug setting on for very long. You can toggle it on/off with questor running FYI.