You can now have faction specific mission XMLs all in the same folder.
The Blockade-Angel Cartel.xml
The Blockade-Blood Raiders.xml
The Blockade-Sansha Nation.xml
The Blockade-Serpentis.xml
The Blockade.xml

if you pulled an angel version of the mission it would use  The Blockade-Angel Cartel.xml, if you pull some oddball faction that we've never seen before or you don't have an xml for it'll use the generic The Blockade.xml.

THIS DOES NOT change existing behavior. All your existing xmls will continue to work. If you don't know what this does for you, ignore it. THose that know what this does, rejoice.