If you do not have room and would have tried to put things in a cargo container created by salvaging a wreck, don't do that... ccp doesn't allow us to put things in those cargo containers anymore. instead create a jetcan...
unfortunately I don't know how to differentiate between a cargo container made by salvaging and a cargo container made by us jettesoning things... if we need to jettison multiple things this is going to be a slow process as we'll need to make a jetcan for every item. (which necessitates waiting 2 min between jetcans)
!!!! There *has* to be a better way to do this. !!!
This fixes a bug though so I am pushing it out as is.