Fix valuedump to still refine if sell is not checked. If neither sell or refine is checked then valuedump will not iterate through the orders.
If you are looking to get a fast 'total' then you will want to uncheck both sell and refine.
If you have refine checked it will iterate through all orders refing things that need to be refined at the end.
if you have sell checked the behavior did not change, if you have both checked the behavior did not change